add check to see if user has completed tutorial #91
